+++ src/test/java/roomescape/acceptance/ReservationTestStep.java
@@ -5,6 +5,7 @@ import static org.assertj.core.api.Assertions.assertThat;
 import io.restassured.RestAssured;
 import io.restassured.http.ContentType;
 import io.restassured.response.Response;
+import java.math.BigDecimal;
 import java.time.LocalDate;
 import java.util.List;
 import roomescape.dto.request.reservation.AdminReservationRequest;
@@ -12,7 +13,8 @@ import roomescape.dto.request.reservation.ReservationRequest;
 
 public class ReservationTestStep {
     public static Long postClientReservation(String token, String date, Long timeId, Long themeId, int expectedHttpCode) {
-        ReservationRequest reservationRequest = new ReservationRequest(LocalDate.parse(date), timeId, themeId, null, null, 1000);
+        ReservationRequest reservationRequest = new ReservationRequest(LocalDate.parse(date), timeId, themeId, null, null,
+                BigDecimal.valueOf(1000));
 
         Response response = RestAssured.given().log().all()
                 .cookies("token", token)
